REFACT :: package Renaming - collection to collector
