+++ src/main/kotlin/com/info/maeumgagym/infrastructure/collector/annotation/DefaultAnnotationCollector.kt
@@ -1,4 +1,4 @@
-package com.info.maeumgagym.infrastructure.collection.annotation
+package com.info.maeumgagym.infrastructure.collector.annotation
 
 import org.springframework.stereotype.Component
 import kotlin.reflect.KClass
